+++ Gemfile
@@ -27,6 +27,7 @@ gem "stimulus-rails"
 # Build JSON APIs with ease [https://github.com/rails/jbuilder]
 gem "jbuilder"
 
+gem "geocoder"
 # Use Redis adapter to run Action Cable in production
 # gem "redis", "~> 4.0"
