+++ app/assets/stylesheets/application.scss
@@ -10,3 +10,4 @@
 // Your CSS partials
 @import "components/index";
 @import "pages/index";
+@import "swiper/swiper-bundle";
