+++ app/assets/stylesheets/components/_caroussel.scss
@@ -1,0 +1,1 @@
+
