+++ app/assets/stylesheets/components/_index.scss
@@ -3,3 +3,4 @@
 @import "avatar";
 @import "form_legend_clear";
 @import "navbar";
+@import "caroussel";
