+++ app/controllers/index.js
@@ -1,0 +1,7 @@
+// app/javascript/controllers/index.js
+
+import { Application } from "stimulus"
+import Carousel from "stimulus-carousel"
+
+const application = Application.start()
+application.register("carousel", Carousel)
