+++ app/views/activities/index.html.erb
@@ -2,11 +2,62 @@
 <div class="container">
 <br>
   <h1 class="title">Perfect Night</h1>
-  <div class="banner"></div>
-  <br>
-    <div style="text-align:center">
-      <p class="description">
-      </p>
-
+    <div class="banner"></div>
+    <br>
+      <div style="text-align:center">
+        <p class="description"></p>
+      </div>
+    </div>
+    <br>
+  <div class="search-bar">
+        <%= form_with url: activities_path, method: :get, class: "d-flex" do %>
+          <%= text_field_tag :query,
+              params[:query],
+              class: "form-control",
+              placeholder: "Recherche"
+            %>
+          <%= submit_tag "Search", class: "btn-purple" %>
+        <% end %>
+  </div>
+    <div id="carouselExampleIndicators" class="carousel slide" data-ride="carousel">
+  <ol class="carousel-indicators">
+    <li data-target="#carouselExampleIndicators" data-slide-to="0" class="active"></li>
+    <li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
+    <li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
+  </ol>
+  <div class="carousel-inner">
+    <div class="carousel-item active">
+      <img class="d-block w-100" src="https://cdn.vox-cdn.com/thumbor/qYhuHHxxjEvqftXbqQk9y7YYSho=/0x145:2000x1192/fit-in/1200x630/cdn.vox-cdn.com/uploads/chorus_asset/file/13199337/2018_09_28_GoldLineBar_003.jpg" alt="First slide">
+        <div class="carousel-caption d-none d-md-block">
+          <h5>Bar dans Paris</h5>
+          <p>Magnifique espace situé au 7ème arrondissement de Paris</p>
+        </div>
+  </div>
+    </div>
+    <div class="carousel-item">
+      <img class="d-block w-100" src="https://wallpapercave.com/wp/wp1929716.jpg" alt="Second slide">
+      <div class="carousel-caption d-none d-md-block">
+          <h5> Bar dans Paris</h5>
+          <p>Magnifique espace situé au 7ème arrondissement de Paris</p>
+          <% activity.name %>
+        </div>
+    </div>
+    <div class="carousel-item">
+      <img class="d-block w-100" src="https://cdn.futura-sciences.com/buildsv6/images/wide1920/9/d/e/9de5f7181a_126842_lobby-1-c-escape-hunt-nantes.jpg" alt="Third slide">
+      <div class="carousel-caption d-none d-md-block">
+          <h5>...</h5>
+          <p>...</p>
+        </div>
+    </div>
+  </div>
+  <a class="carousel-control-prev" href="#carouselExampleIndicators" role="button" data-slide="prev">
+    <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+    <span class="sr-only">Previous</span>
+  </a>
+  <a class="carousel-control-next" href="#carouselExampleIndicators" role="button" data-slide="next">
+    <span class="carousel-control-next-icon" aria-hidden="true"></span>
+    <span class="sr-only">Next</span>
+  </a>
+</div>
   <% end %>
 </div>
