+++ db/migrate/20230206111119_add_coordinates_to_flats.rb
@@ -1,0 +1,6 @@
+class AddCoordinatesToFlats < ActiveRecord::Migration[7.0]
+  def change
+    add_column :flats, :latitude, :float
+    add_column :flats, :longitude, :float
+  end
+end
