+++ package.json
@@ -6,6 +6,7 @@
     "@hotwired/turbo-rails": "^7.2.5",
     "@popperjs/core": "^2.11.6",
     "bootstrap": "^5.2.3",
+    "stimulus-carousel": "^5.0.1",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1"
   },
